fix: move wallet types to separate file
